backend_tapiceria_api/requirements.txt
asgiref==3.8.1
attrs==25.3.0
cloudinary==1.41.0
dj-database-url==3.0.1
Django==4.2.21
django-cors-headers==4.7.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
packaging==25.0
psycopg==3.3.2
psycopg-binary==3.3.2
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
whitenoise==6.11.0
django-cloudinary-storage==0.3.0

# --- DEPENDENCIAS PARA PDF ---
# Actualizado a versiones estables y compatibles
reportlab==4.2.5
pillow==11.0.0
xhtml2pdf==0.2.17
requests==2.32.3

# --- SERVIDOR WEB ---
gunicorn==23.0.0